change false embedder to utilise minimum length sequence when reconstructing fake encoder output

+++ molecules/transformer.py
@@ -832,7 +832,7 @@ class FalseEmbeddingsNonTD():
         :param d_emb: dimensionality of false embeddings
         '''
         NUM_LAYERS = 0
-        latent_len = np.ceil(d_latent/d_emb)*d_emb
+        latent_len = int(np.ceil(d_latent/d_emb)*d_emb)
         self.init_layer = Dense(d_emb * latent_len, input_shape=(d_latent,))
 
         # self.init_layer = TimeDistributed(Dense(d_emb, input_shape=(1,)))
